Move `DeviceAlignment`
