+++ vulkano/src/buffer/allocator.rs
@@ -13,9 +13,12 @@ use super::{Buffer, BufferContents, BufferError, BufferMemory, BufferUsage, Subb
 use crate::{
     buffer::BufferAllocateInfo,
     device::{Device, DeviceOwned},
-    memory::allocator::{
-        align_up, AllocationCreationError, DeviceAlignment, DeviceLayout, MemoryAllocator,
-        MemoryUsage, StandardMemoryAllocator,
+    memory::{
+        allocator::{
+            align_up, AllocationCreationError, DeviceLayout, MemoryAllocator, MemoryUsage,
+            StandardMemoryAllocator,
+        },
+        DeviceAlignment,
     },
     DeviceSize, NonZeroDeviceSize,
 };
